manejo de errores

+++ server.py
@@ -4,54 +4,76 @@ import os
 import asyncio
 import argparse
 import sys
+from urllib import request
 import db
 from string import Template
+from parceo import Parceo
 
-filein= open(os.getcwd() + '/index.html')
-src=Template(filein.read())
 
 cantidad_sensores=2
 
 async def handle(reader, writer):
     
     data = await reader.read(100000)                                #Datos que recibe el servidor
-    print(data)
+    #print(data)
     
     if  data.decode()[0:6] == 'sensor':                             #Solicitud del sensor 
         print ('Dato recibido')
         db.insert(data.decode())
     
 
-    else:                                                           #Solicitud web
-        ult_mediciones = db.select_valor(cantidad_sensores)
-        D={}
-        for num in range(cantidad_sensores):
-            D['sensor'+str(num+1)]= ult_mediciones[num][0]
-            D['tipo'+str(num+1)]= ult_mediciones[num][1]
-            D['medicion'+str(num+1)]= ult_mediciones[num][2]
-            D['fecha'+str(num+1)]= ult_mediciones[num][3]
-        v_web= src.substitute(D)                                           
+    else:                                                                   #Solicitud web
+        
+        consulta=Parceo.parcear(data)
+        print(consulta)
+        
+        if consulta[0] == "GET":                                                                #Respuesta a  metodo GET
+            if os.path.isfile(consulta[1].replace('/','')) == True or consulta[1]== '/':        #Respuesta si existe el archivo o /
+                filein= open(os.getcwd() + '/index.html')
+                src=Template(filein.read())
+
+                ult_mediciones = db.select_valor(cantidad_sensores)
+                D={}
+                for num in range(cantidad_sensores):
+                    D['sensor'+str(num+1)]= ult_mediciones[num][0]
+                    D['tipo'+str(num+1)]= ult_mediciones[num][1]
+                    D['medicion'+str(num+1)]= ult_mediciones[num][2]
+                    D['fecha'+str(num+1)]= ult_mediciones[num][3]
+                v_web= src.substitute(D)                                           
+                path = os.getcwd() + '/filein.html'
+                fd = open(path,'w')
+                fd.writelines(v_web)
+                fd.close()
+                fd2=os.open(path, os.O_RDONLY)
+                body = os.read(fd2,os.stat(path).st_size)
+                os.close(fd2)
+                respuesta= '200 OK'
+                header = bytearray("HTTP/1.1 " + respuesta + "\r\nContent-type:" + 'text/html' 
+                            +"\r\nContent-length:" + str(len(body)) + "\r\n\r\n",'utf8')
+
+            elif os.path.isfile(consulta[1].replace('/','')) == False:                          #Respuesta si no existe el archivo
+                path = os.getcwd() + '/error404.html'
+                fd2=os.open(path, os.O_RDONLY)
+                body = os.read(fd2,os.stat(path).st_size)
+                os.close(fd2)
+                respuesta= '404 Not Found'
+                header = bytearray("HTTP/1.1 " + respuesta + "\r\nContent-type:" + 'text/html' 
+                            +"\r\nContent-length:" + str(len(body)) + "\r\n\r\n",'utf8')
+                
+                
+        else:                                                                                   #Respuesta si el metodo no esta permitido
+            path = os.getcwd() + '/error405.html'
+            fd2=os.open(path, os.O_RDONLY)
+            body = os.read(fd2,os.stat(path).st_size)
+            os.close(fd2)
+            respuesta= '405 Method Not Allowed'
+            header = bytearray("HTTP/1.1 " + respuesta + "\r\nContent-type:" + 'text/html' 
+                        +"\r\nContent-length:" + str(len(body)) + "\r\n\r\n",'utf8')
 
 
-        path = os.getcwd() + '/filein.html'
-        fd = open(path,'w')
-        fd.writelines(v_web)
-        fd.close()
-
-        fd2=os.open(path, os.O_RDONLY)
-        body = os.read(fd2,os.stat(path).st_size)
-        #print(body)
-        os.close(fd2)
-        respuesta= '200 OK'
-        header = bytearray("HTTP/1.1 " + respuesta + "\r\nContent-type:" + 'text/html' 
-                    +"\r\nContent-length:" + str(len(body)) + "\r\n\r\n",'utf8')
-
-        writer.write(header)
-        writer.write(body)
-        #await writer.drain()
-        writer.close()
-
-    
+    writer.write(header)                                                                        #Respondemos con la cabecera
+    writer.write(body)                                                                          #Respondemos con el body
+    writer.close()
 
 
 async def main():
